Fix Windows MSVC warnings and SlimeMold indentation so all CI jobs build clean
- SlimeMold: split single-line if chains that tripped -Werror=misleading-indentation
- C4996: add _CRT_SECURE_NO_WARNINGS for MSVC (getenv is standard C)
- C4477: use %hs for narrow strings in wide swprintf labels (6 examples)
- C4267: explicit size_t->int cast in Tweening
- C4244: explicit int->float casts at 68 narrowing sites across examples/games
  (burst/viewportToWorld/setSize/Rect/emit args, float member assigns)
- C4458: suppress the deliberate getRenderer() idiom (/wd4458, documented) and
  rename genuine shadowing (assetManager, buttons, deltaTime, tileSize,
  drawNinePatch/drawTiledSprite/drawUI renderer params)

+++ Examples/2DViewportToWorld/main.cpp
@@ -88,8 +88,7 @@ public:
         
         // Get mouse position and convert to world space
         int mx, my;
-        SDL_GetMouseState(&mx, &my);
-        mouseWorldPos = camera->viewportToWorld(mx, my);
+        SDL_GetMouseState(&mx, &my);          mouseWorldPos = camera->viewportToWorld((float)mx, (float)my);
     }
 
     void renderGame() override {
